Fix load logs in loaders to correct output loader (and child loaders) logs
The load_log was stored as an class variable
instead of an instance variable, so all instances
of the class were sharing the same variable. It
doesn't seem possible to overwrite this variable
so it made sense to switch to an variable for each
instance.

The BaseLoader class initialiser now sets the load
log variable. If none is given in a loaders
__init__() then a new list is created (passed through
list() to avoid Python using previous load logs in
memory). If a load log is given, it saves a reference
to this list.

All loaders which call another loader pass the reference
to their load log for the loader to append to.

This is related to #46.