Add breadcrumbs to topic pages
I have only added breadcrumbs to ages that
will be viewed by users. We are still
unsure how if the resources app can be
viewed independently, and if programming
difficulty pages will be viewed, so these
were skipped.

The approach of breadcrumbs is that templates
inherit from the page above them in the
breadcrumb trail, instead of base.html. This
adds computational overhead in generating
templates, but this appears to be very
minimal.

It seems like this approach is suitable,
unless we try to improve template rendering
speed (but usually other bottlenecks are
targeted first).