Update to card class for homepage and topic pages
Issues I've discovered using cards is problems
with setting widths with different view sizes
and also margins when there are enough items
to wrap into a new row.

Will possibly need to create something from
scratch using flexbox, much like buttons on
CSFG homepage. The appearance should still
be the same.

The current code is fine as this is just a
prototype.