Only show DBT in docker-compose mode, not on localhost
This was a very odd bug, and I don't completely
understand why the test suite is breaking because of
this.

Initially I had hardcoded the debug toolbar to always
show, and I now know it's because the internal
IP wasn't found in the list (the cookiecutter code
expected the tail to be '1' but I've had it running
as '4'. I could switch the function to match the
first section of the IP, but since this code is only
in the local settings code it should be fine as
hardcoded as TRUE.

The local dev mode runs with DEBUG as TRUE but the
test suite runs DEBUG with FALSE for it's tests,
and it seems to be trying to find the debug toolbar
URLs but these are hidden behind a DEBUG check.

I can't find anyone else with this issue online.

The best option would be to switch the tests to
run with the test suite configuration file, as this
runs with DEBUG as FALSE and never import the
django debug toolbar.