Modify custom loaders to populate translated field content (#615)
The presence of a content markdown file in the directory tree of a
given language is used to determine that a translation is available
for that language. When present, the markdown file is parsed and
loaded into the respective field (provided by django-modeltranslation).
The available languages field is also updated to reflect this.

This required some restructuring regarding the way paths are stored
in the loader class, because it needs to be possible to retrieve the
content file from any language's directory tree.