Add model for classroom resources, defined in their own yaml file.
This was required for translation of the resource strings, using
vinaigrette (which didn't support translation of an ArrayField),
relevant to #612.

Lessons now refer to a classroom resources that is defined in
the classroom-resources yaml file.

Authors were consulted about changes to the way classroom resources
are written - it was determined these need a bit of an overhaul
throughout all content to be more in line with the intended purpose.
This commit provides a starting point, but will need to be reviewed
by authors later in the process.