Add curriculum areas filter
Currently the given filter is provided as a QuerySet, but the search
index saves the curriculum areas for objects (currently only
curriculum integrations) as a list of primary keys, but stored as
strings rather than integers. This may be due to the way data is
stored.

Because of this, the logic below must covert the QuerySet of the
filter into a list of primary key strings.

I also could not get Django/Haystack to filter by checking if any
item in the filter was any item in each object in the index.
Something similar to the following:
search_query_set = search_query_set.filter(
   curriculum_areas__in=self.cleaned_data["curriculum_areas"]
)