Update classic content file as it will be only used for searching
This is because once a URL regular expression matches,
it must return a view. Therefore we can't take the initial
strategy of if an object doesn't exist to then return
to the URL routes. This new approach will have duplicate
data but results in no database queries when trying to find
a URL match.